små ændringer i shoppen + ændringer fra int(price) til double igennem hele systemet

+++ src/main/java/data/CompleteCupCake.java
@@ -11,17 +11,17 @@ package data;
  */
 public class CompleteCupCake {
     private String name;
-    private int price;
+    private double price;
 
     public String getName() {
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public CompleteCupCake(String name, int price) {
+    public CompleteCupCake(String name, double price) {
         this.name = name;
         this.price = price;
     }
